packages/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlmodel==0.0.22
pydantic==2.10.3
pydantic-settings==2.6.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
better-exceptions==0.3.3
python-multipart==0.0.20
alembic==1.14.0
asyncpg==0.30.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
pytest==8.3.4
httpx==0.27.2